Fix UBSAN implicit conversion warning in test/test_shared_ng.h.

+++ test/test_shared_ng.h
@@ -9,12 +9,12 @@ static inline int deflate_prime_32(PREFIX3(stream) *stream, uint32_t value) {
     int err;
 
 #ifndef TEST_STOCK_ZLIB
-    err = PREFIX(deflatePrime)(stream, 32, value);
+    err = PREFIX(deflatePrime)(stream, 32, (int32_t)value);
 #else
     /* zlib's deflatePrime() takes at most 16 bits */
-    err = PREFIX(deflatePrime)(stream, 16, value & 0xffff);
+    err = PREFIX(deflatePrime)(stream, 16, (int32_t)(value & 0xffff));
     if (err != Z_OK) return err;
-    err = PREFIX(deflatePrime)(stream, 16, value >> 16);
+    err = PREFIX(deflatePrime)(stream, 16, (int32_t)(value >> 16));
 #endif
 
     return err;
